move methods from ModifiedReplay to Comparer
