+++ circleguard/loadable.py
@@ -550,9 +550,8 @@ class Replay(Loadable):
 
         Returns
         -------
-        array[int], array[[float]], array[int]
-            A list of tuples of (t, x, y, keys) for each event
-            in the replay data.
+        ndarray[int], ndarray[[float]], ndarray[int]
+            The arrays of time, x and y, and pressed keys.
         """
         block = list(zip(*[(e.time_since_previous_action, e.x, e.y, e.keys_pressed) for e in self.replay_data]))
 
@@ -727,317 +726,3 @@ class ReplayPath(Replay):
         Replay.__init__(self, loaded.timestamp, map_id, loaded.player_name, user_id, ModCombination(loaded.mod_combination),
                         loaded.replay_id, loaded.play_data, self.weight)
         self.log.log(TRACE, "Finished loading %s", self)
-
-
-class ReplayModified(Replay):
-    """
-    A :class:`~.Replay` created by modifying or copying another :class:`~.Replay`.
-
-    Parameters
-    ----------
-    timestamp: :class:`datetime.datetime`
-        When this replay was played.
-    map_id: int
-        The id of the map the replay was played on, or 0 if
-        unknown or on an unsubmitted map.
-    user_id: int
-        The id of the player who played the replay, or 0 if unknown
-        (if the player is restricted, for instance). Note that if the
-        user id is known, even if the user is restricted, it should still be
-        given instead of 0.
-    username: str
-        The username of the player who played the replay.
-    mods: :class:`~.enums.ModCombination`
-        The mods the replay was played with.
-    replay_id: int
-        The id of the replay, or 0 if the replay is unsubmitted.
-    replay_data: list[:class:`~circleparse.Replay.ReplayEvent`]
-        A list of :class:`~circleparse.Replay.ReplayEvent` objects, representing
-        the actual data of the replay. If the replay could not be loaded, this
-        should be ``None``.
-    weight: :class:`~.enums.RatelimitWeight`
-        How much it 'costs' to load this replay from the api.
-    """
-
-    def __init__(self, timestamp, map_id, username, user_id, mods, replay_id, replay_data, weight):
-        Replay.__init__(self, timestamp, map_id, username, user_id, mods, replay_id, replay_data, weight)
-        self.txyk = None
-
-    def load(self, loader):
-        """
-        Do nothing as this :class:`~.ReplayModified` is always already loaded.
-        """
-        pass
-
-    def interpolate_to(self, timestamps):
-        """
-        Interpolate coordinate data in this :class:`~.ReplayModified` to the given timestamps.
-        Drops excess timestamps and creates data at missing timestamps.
-
-        Parameters
-        ----------
-        timestamps: list(int)
-            The timestamps to interpolate to.
-
-        Notes
-        -----
-        This method is in-place.
-        """
-        prev_err = np.seterr(all="ignore")
-
-        txyks = []
-
-        self.as_list_with_timestamps()
-
-        i = 1
-        t_end = self.txyk[-1][0]
-
-        for t in timestamps:
-            if t > t_end:
-                break
-
-            while self.txyk[i][0] < t:
-                i += 1
-
-            r = (t - self.txyk[i - 1][0]) / (self.txyk[i][0] - self.txyk[i - 1][0])
-
-            if np.isnan(r) or np.isinf(r):
-                continue
-
-            x = r * self.txyk[i][1] + (1 - r) * self.txyk[i - 1][1]
-            y = r * self.txyk[i][2] + (1 - r) * self.txyk[i - 1][2]
-            k = self.txyk[i - 1]
-
-            txyks += [[t, x, y, k]]
-
-        self.txyk = txyks
-
-        np.seterr(**prev_err)
-
-        return self
-
-    def shift(self, dx, dy, dt):
-        """
-        Shift the data in this :class:`~.ReplayModified` by the specified offsets.
-
-        Parameters
-        ----------
-        dx: float
-            The offset for the x coordinate.
-        dy: float
-            The offset for the y coordinate.
-        dt: int
-            The number of ticks the timestamps should be shifted.
-
-        Notes
-        -----
-        This method is in-place.
-        """
-        txyks = []
-
-        self.as_list_with_timestamps()
-
-        for t, x, y, k in self.txyk:
-            txyks += [[t + dt, x + dx, y + dy, k]]
-
-        self.txyk = txyks
-
-        return self
-
-    def as_list_with_timestamps(self):
-        """
-        Gets this replay's play data as a list of tuples of absolute time,
-        x, y, and pressed keys for each event in the data.
-        Saves the data which is returned on consequent calls.
-        This data reflects changes made by other methods.
-
-        Returns
-        -------
-        list[tuple(int, float, float, something)]
-            A list of tuples of (t, x, y, keys) for each event
-            in the replay data.
-        """
-        if not self.txyk:
-            self.txyk = Replay.as_list_with_timestamps(self)
-
-        return self.txyk
-
-    @staticmethod
-    def copy(replay):
-        """
-        Copies the :class:`~.Replay`
-
-        Parameters
-        ----------
-        replay: :class:`~.Replay`
-            The replay to copy.
-
-        Returns
-        -------
-        :class:`~.ReplayModified`
-            The copy as a :class:`~.ReplayModified`.
-        """
-
-        r = ReplayModified(replay.timestamp, replay.map_id, replay.username,
-                      replay.user_id, replay.mods, replay.replay_id, replay.replay_data, replay.weight)
-        r.txyk = replay.as_list_with_timestamps()
-        return r
-
-    @staticmethod
-    def filter_single(replay):
-        """
-        Filters all timestamps with invalid coordinates.
-
-        Parameters
-        ----------
-        replay: :class:`~.ReplayModified`
-            The replay to filter.
-
-        Returns
-        -------
-        :class:`~.ReplayModified`
-            The replay with the invalid coordinates removed.
-
-        Notes
-        -----
-        Coordinates are invalid if they are not in the range (0, 512) for the x
-        coordinate and (0, 384) for the y coordinate.
-        This method is in-place.
-        """
-        data = replay.as_list_with_timestamps()
-
-        def is_valid(d):
-            return 0 <= d[1] <= 512 and 0 <= d[2] <= 384
-
-        txyk = [d1 for (d0, d1) in zip(data, data[1:]) if is_valid(d0) and is_valid(d1)]
-
-        replay.loaded = True
-        replay.txyk = txyk
-
-        return replay
-
-    @staticmethod
-    def align_clocks(clocks):
-        """
-        Selects suitable timestamps to which all the clocks of the replays
-        could be interpolated.
-
-        The timestamps are chosen to maximize frequency while
-        also guaranteeing each replay has at least one timestamp in between
-        two successive timestamps in the output.
-
-        Parameters
-        ----------
-        clocks: list(list(int))
-            The list of lists of timestamps the interpolation timestamps
-            should be selected from.
-
-        Returns
-        -------
-        list(int)
-            The timestamps maximizing frequency and satisfying constraints.
-        """
-        n = len(clocks)
-        indices = [0] * n
-
-        output = []
-
-        def move_to(i, value):
-            while clocks[i][indices[i] + 1] <= value:
-                indices[i] += 1
-
-                if indices[i] + 1 == len(clocks[i]):
-                    return True
-
-            return False
-
-        while True:
-            last = 0
-
-            for i in range(n):
-                value = clocks[i][indices[i] + 1]
-
-                if value > last:
-                    last = value
-
-            for i in range(n):
-                if move_to(i, last):
-                    return output
-
-            output += [last]
-
-    @staticmethod
-    def align_coordinates(replays):
-        """
-        Shifts the :class:`~.ReplayModified`s in replays so that their
-        mean coordinates over time coincide.
-
-        Parameters
-        ----------
-        replays: list(:class:`~.ReplayModified`)
-            The replays to align.
-
-        Returns
-        -------
-        list(:class:`~.ReplayModified`)
-            The shifted replays.
-
-        Notes
-        -----
-        The first replay in this set is left in place, and all other replays
-        are shifted toward its mean.
-        This method is in-place.
-        """
-        rs = []
-
-        for replay in replays:
-            replay.as_list_with_timestamps()
-
-            xs = []
-            ys = []
-
-            for _, x, y, _ in replay.txyk:
-                xs += [x]
-                ys += [y]
-
-            rs += [(np.mean(xs), np.mean(ys))]
-
-        replays = replays[:1] + [replay.shift(rs[0][0] - r[0], rs[0][1] - r[1], 0) for (replay, r)
-                                in zip(replays[1:], rs[1:])]
-
-        return replays
-
-    @staticmethod
-    def clean_set(replays, mode):
-        """
-        Cleans the :class:`~.ReplayModified`s in replays using the methods specified in the mode.
-
-        Parameters
-        ----------
-        replays: list(:class:`~.ReplayModified`)
-            The replays to clean.
-        mode: :class:`~.CleanMode`
-            The mode specifying the used methods.
-
-        Returns
-        -------
-        list(:class:`~.ReplayModified`)
-            The cleaned :class:`~.ReplayModified`s
-        """
-
-        replays = [ReplayModified.copy(r) for r in replays]
-
-        if CleanMode.VALIDATE in mode:
-            replays = [ReplayModified.filter_single(replay) for replay in replays]
-
-        if CleanMode.SYNCHRONIZE in mode:
-            timestamps = [[txyk[0] for txyk in replay.as_list_with_timestamps()] for replay in replays]
-
-            timestamps = ReplayModified.align_clocks(timestamps)
-
-            replays = [replay.interpolate_to(timestamps) for replay in replays]
-
-        if CleanMode.ALIGN in mode:
-            replays = ReplayModified.align_coordinates(replays)
-
-        return replays
